Moved fetchFromGist to seperate file
