+++ src/index.js
@@ -4,6 +4,8 @@ import { Router, Route, IndexRedirect, browserHistory } from 'react-router';
 
 import NotebookPreview from 'notebook-preview';
 
+import { fetchFromGist } from './fetchers';
+
 const includes = require('lodash.includes');
 
 const gistIDs = [
@@ -28,26 +30,7 @@ const Main = React.createClass({
 const Notebook = React.createClass({
   componentDidMount: function() {
     this.setState({
-      nbJSON: fetch(`https://api.github.com/gists/${this.props.params.gistId}`)
-        .then((data) => data.json())
-        .then((ghResponse) => {
-          for (var file in ghResponse.files) {
-            if (includes(file, '.ipynb')) {
-              const fileResponse = ghResponse.files[file];
-              if (fileResponse.truncated) {
-                return fetch(fileResponse.raw_url)
-                          .then((resp) => resp.json())
-              }
-
-              const nbString = fileResponse.content;
-              const nbJSON = JSON.parse(nbString);
-              return nbJSON;
-            }
-          }
-        })
-        .then((nbJSON) => {
-          return nbJSON;
-        })
+      nbJSON: fetchFromGist(this.props.params.gistId)
     });
   },
   render: function() {
